app_whats/requirements.txt
aiofiles==24.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
cookiecutter==2.6.0
customtkinter==5.2.2
darkdetect==0.8.0
debugpy==1.8.2
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.112.1
Flask==3.0.3
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
frontend==0.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
keyboard==0.13.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
nest-asyncio==1.6.0
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyPDF2==3.0.1
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.3
pywinauto==0.6.8
PyYAML==6.0.2
pyzmq==26.0.3
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==13.7.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.38.2
text-unidecode==1.3
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.4
wikipedia==1.4.0
wsproto==1.2.0